work in progress 11/07/2019

+++ BL/BookBL.cs
@@ -2,6 +2,7 @@
 using DAL;
 using DAL.Interface;
 using Entities;
+using System;
 using System.Collections.Generic;
 
 namespace BL
@@ -12,27 +13,66 @@ namespace BL
 
         public void Add(Book book)
         {
-            bookDao.Add(book);
+            if (book.isCorrectState())
+            {
+                bookDao.Add(book);
+            }
+            else
+            {
+                throw new ArgumentException("Error, Added book in incorrec state (empty fields)");
+            }
+            
         }
 
         public void Remove(int ID)
         {
-            bookDao.Remove(ID);
+            if (ID > 0)
+            {
+                bookDao.Remove(ID);
+            }
+            else
+            {
+                throw new ArgumentException("ID is negative or zero");
+            }
+            
         }
 
         public void Update(Book book)
         {
-            bookDao.Update(book);
+            if (book.isCorrectState())
+            {
+                bookDao.Update(book);
+            }
+            else
+            {
+                throw new ArgumentException("Error, updated book in incorrec state (empty fields)");
+            }
+            
         }
 
         public IEnumerable<Book> GetBooksByTitle(string title)
         {
-            return bookDao.GetBooksByTitle(title);
+            if (!string.IsNullOrEmpty(title))
+            {
+                return bookDao.GetBooksByTitle(title);
+            }
+            else
+            {
+                throw new ArgumentException("Value empty or null -> title");
+            }
+            
         }
 
         public IEnumerable<Book> GetBooksByAuthor(string author)
         {
-            return bookDao.GetBooksByAuthor(author);
+            if (!string.IsNullOrEmpty(author))
+            {
+                return bookDao.GetBooksByAuthor(author);
+            }
+            else
+            {
+                throw new ArgumentException("Value empty or null -> author");
+            }
         }
 
         public IEnumerable<Book> GetAllBooks()
@@ -42,7 +82,15 @@ namespace BL
 
         public IEnumerable<UserBook> GetBooksByID(int Reader_ID)
         {
-            return bookDao.GetBooksByID(Reader_ID);
+            if (Reader_ID > 0)
+            {
+                return bookDao.GetBooksByID(Reader_ID);
+            }
+            else
+            {
+                throw new ArgumentException("Reader_ID is negative or zero");
+            }
+            
         }
     }
 }
